refactor: more handshake more handshake logic to handshaker
